requirements.txt
django-heroku==0.3.1
gunicorn==19.9.0
crython==0.2.0
django==2.2.2
djangorestframework==3.9.4
django-extensions==2.1.7
requests==2.21.0
django-environ==0.4.5
